# the-localhosts-main/api/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "SnapRx REST API Documentation",
    "version": "1.0.0",
    "title": "SnapRX REST API Documentation",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/users/signup": {
      "post": {
        "tags": [
          "User Resources"
        ],
        "summary": "Account Sign Up",
        "requestBody": {
          "description": "Sign Up Request object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signupRequest"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/signupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/statusMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "User Resources"
        ],
        "summary": "Account login",
        "requestBody": {
          "description": "Login Request object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginRequest"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/statusMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "medications/search": {
      "post": {
        "tags": [
          "Medication Resources"
        ],
        "summary": "Search medications",
        "requestBody": {
          "description": "Search Medications Request object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postSearchMedicationsRequest"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/postSearchMedicationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/statusMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "medications/classify-by-description": {
      "post": {
        "tags": [
          "Medication Resources"
        ],
        "summary": "Classify medication by physical description",
        "requestBody": {
          "description": "Classify Medication by Description Request object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postClassifyMedicationByDescriptionRequest"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/postClassifyMedicationByDescriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/statusMessageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "signupRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "required": true,
            "description": "New user email"
          },
          "password": {
            "type": "string",
            "required": true,
            "description": "New user password"
          },
          "firstName": {
            "type": "string",
            "required": true,
            "description": "New user first name"
          },
          "lastName": {
            "type": "string",
            "required": true,
            "description": "New user last name"
          }
        }
      },
      "signupResponse": {
        "properties": {
          "userId": {
            "type": "integer",
            "required": true,
            "description": "ID of new user"
          },
          "token": {
            "type": "string",
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}",
            "required": true,
            "description": "Session token"
          },
          "message": {
            "type": "string",
            "required": false,
            "description": "Status message"
          },
          "errors": {
            "type": "array",
            "required": false,
            "description": "Array of errors"
          }
        }
      },
      "loginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "required": true,
            "description": "User email"
          },
          "password": {
            "type": "string",
            "required": true,
            "description": "User password"
          }
        }
      },
      "loginResponse": {
        "properties": {
          "userId": {
            "type": "integer",
            "required": true,
            "description": "ID of user"
          },
          "token": {
            "type": "string",
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}",
            "required": true,
            "description": "Session token"
          },
          "message": {
            "type": "string",
            "required": false,
            "description": "Status message"
          },
          "errors": {
            "type": "array",
            "required": false,
            "description": "Array of errors"
          }
        }
      },
...
      "postSearchMedicationsRequest": {
        "properties": {
          "query": {
            "type": "string",
            "required": true,
            "description": "Search query for medication"
          }
        }
      },
      "postSearchMedicationsResponse": {
        "properties": {
          "results": {
            "type": "array",
            "required": true,
            "description": "Array of medication search results",
            "items": {
              "type": "object",
              "properties": {
                "medId": {
                  "type": "integer",
                  "required": true,
                  "description": "ID of medication"
                },
                "medName": {
                  "type": "string",
                  "required": true,
                  "description": "Medication name"
                },
                "medDetails": {
                  "type": "string",
                  "required": true,
                  "description": "Medication details string"
                },
                "brandId": {
                  "type": "integer",
                  "required": false,
                  "description": "ID of manufacturing brand"
                },
                "price": {
                  "type": "number",
                  "required": false,
                  "description": "Price of medication"
                },
                "rxString": {
                  "type": "string",
                  "required": false,
                  "description": "Full prescription string"
                },
                "color": {
                  "type": "string",
                  "required": false,
                  "description": "Color of medication"
                },
                "imprintBack": {
                  "type": "string",
                  "required": false,
                  "description": "Imprint on back of medication"
                },
                "imprintFront": {
                  "type": "string",
                  "required": false,
                  "description": "Imprint on front of medication"
                },
                "shape": {
                  "type": "string",
                  "required": false,
                  "description": "Shape of medication"
                },
                "size": {
                  "type": "number",
                  "required": false,
                  "description": "Size of medication"
                }
              }
            }
          },
          "message": {
            "type": "string",
            "required": false,
            "description": "Status message"
          },
          "errors": {
            "type": "array",
            "required": false,
            "description": "Array of errors"
          }
        }
      },
...
      "postClassifyMedicationByDescriptionRequest": {
        "properties": {
          "color": {
            "type": "string",
            "enum": [
              "PINK",
              "ORANGE",
              "GREEN",
              "WHITE",
              "YELLOW",
              "BLUE",
              "RED",
              "BROWN",
              "PURPLE",
              "GRAY",
              "TURQUOISE",
              "BLACK"
            ],
            "required": false,
            "description": "Color of medication"
          },
          "color2": {
            "type": "string",
            "enum": [
              "PINK",
              "ORANGE",
              "GREEN",
              "WHITE",
              "YELLOW",
              "BLUE",
              "RED",
              "BROWN",
              "PURPLE",
              "GRAY",
              "TURQUOISE",
              "BLACK"
            ],
            "required": false,
            "description": "Secondary color of medication"
          },
          "imprintBack": {
            "type": "string",
            "required": false,
            "description": "Imprint on back of medication"
          },
          "imprintFront": {
            "type": "string",
            "required": false,
            "description": "Imprint on front of medication"
          },
          "shape": {
            "type": "string",
            "enum": [
              "CAPSULE",
              "ROUND",
              "PENTAGON",
              "OVAL",
              "RECTANGLE",
              "DIAMOND",
              "OCTAGON",
              "HEXAGON",
              "BULLET",
              "SQUARE",
              "FREEFORM",
              "TRIANGLE",
              "SEMI-CIRCLE",
              "TEAR",
              "DOUBLE CIRCLE",
              "TRAPEZOID",
              "CLOVER"
            ],
            "required": false,
            "description": "Shape of medication"
          },
          "size": {
            "type": "number",
            "required": false,
            "description": "Size of medication"
          }
        }
      },
      "postClassifyMedicationByDescriptionResponse": {
        "properties": {
          "results": {
            "type": "array",
            "required": true,
            "description": "Array of classifier results",
            "items": {
              "type": "object",
              "properties": {
                "medId": {
                  "type": "integer",
                  "required": true,
                  "description": "ID of medication"
                },
                "medName": {
                  "type": "string",
                  "required": true,
                  "description": "Medication name"
                },
                "medDetails": {
                  "type": "string",
                  "required": true,
                  "description": "Medication details string"
                },
                "brandId": {
                  "type": "integer",
                  "required": false,
                  "description": "ID of manufacturing brand"
                },
                "price": {
                  "type": "number",
                  "required": false,
                  "description": "Price of medication"
                },
                "rxString": {
                  "type": "string",
                  "required": false,
                  "description": "Full prescription string"
                },
                "color": {
                  "type": "string",
                  "required": false,
                  "description": "Color of medication"
                },
                "imprintBack": {
                  "type": "string",
                  "required": false,
                  "description": "Imprint on back of medication"
                },
                "imprintFront": {
                  "type": "string",
                  "required": false,
                  "description": "Imprint on front of medication"
                },
                "shape": {
                  "type": "string",
                  "required": false,
                  "description": "Shape of medication"
                },
                "size": {
                  "type": "number",
                  "required": false,
                  "description": "Size of medication"
                }
              }
            }
          },
          "message": {
            "type": "string",
            "required": false,
            "description": "Status message"
          },
          "errors": {
            "type": "array",
            "required": false,
            "description": "Array of errors"
          }
        }
      },
      "badRequestResponse": {
        "properties": {
          "errors": {
            "type": "array",
            "required": false,
            "description": "Array of errors"
          }
        }
      },
      "statusMessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "required": false,
            "description": "Status message"
          }
        }
      }
    }
  }
}